Simple fix I forgot to commit when I changed the year field from an int to a string
git-svn-id: http://svn.mythtv.org/svn/trunk@8410 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythplugins/mythgame/mythgame/rominfo.cpp
@@ -132,7 +132,7 @@ void RomInfo::setField(QString field, QString data)
     else if (field == "genre")
         genre = data;
     else if (field == "year")
-        year = data.toInt();
+        year = data;
     else if (field == "favorite")
         favorite = data.toInt();
     else if (field == "rompath")
